Make scripts Django management commands
